Health and Audio Fixes

+++ Assets/Scripts/PlayerScripts/HealthBar.cs
@@ -7,15 +7,13 @@ public class HealthBar : MonoBehaviour
 {
     // Start is called before the first frame update
     public Slider slider;
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health;
     }
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
-
-   
 }
